(textureX and textureY are now double)

+++ Entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.h"
 //X and Y are Spawn Positions, textureX and textureY are height and width of sprites
-Entity::Entity(int x, int y,int tolerance, string texturePath, string name, int textureX, int textureY, int sprites) : pos(x, y), tolerance(tolerance),textureX(textureX),textureY(textureY), name(name),sprites(sprites),x(0),reverse(false) {
+Entity::Entity(int x, int y,int tolerance, string texturePath, string name, double textureX,double textureY, int sprites) : pos(x, y), tolerance(tolerance),textureX(textureX),textureY(textureY), name(name),sprites(sprites),x(0),reverse(false) {
 		texture.loadFromFile(texturePath);
 		sprite.setTexture(texture);
 		sprite.setTextureRect(IntRect(0, 0, textureX, textureY));
